pythonapp/requirements.txt
Django==1.11.2
wheel==0.29.0
gunicorn==19.9.0
psycopg2==2.7.1
flask>=1.0.0
flask_sqlalchemy==2.3.2
pymysql==0.9.2
flask_pymongo>=2.0.0
oauth2client==4.1.2
pymongo==3.7.2
six==1.11.0
itsdangerous==0.24
jinja2>=2.10.1
MarkupSafe>=1.0
pytz==2017.2
requests>=2.18.0
werkzeug>=0.15.0